solver and writer class

+++ Module_1_solution/src/writer.hpp
@@ -3,6 +3,8 @@
 
 #include "array_2d.hpp"
 #include <string> //std::string
+#include <iostream>
+#include <fstream>
 
 class ArrayWriter{
 	
@@ -12,12 +14,14 @@ public:
     
 
     void write_to_file(std::string o_file_path, const Array2d& array);
-    void write_mesh_to_file(std::string o_file_path, double Lx, double Ly, int nx, int ny);
+    void write_mesh_to_file(std::string o_file_path, double Lx, double Ly, size_t nx, size_t ny);
     
 
 protected:
  
 private:
+    
+
           	
 };
